Tiny improvements using long masks instead of word masks

+++ snowflake.h
@@ -26,9 +26,9 @@ void display_flake_sprite_opt(unsigned short* video_ptr,
                               unsigned short* backsnow_ptr,
                               unsigned short* background_ptr,
                               unsigned short* flake_pic,
-                              unsigned short* flake_mask,
-                              unsigned short* background_mask,
-                              unsigned short* foreground_mask,
+                              unsigned long* flake_mask,
+                              unsigned long* background_mask,
+                              unsigned long* foreground_mask,
                               unsigned short foreground_flake_flag);
 
 #endif
